+++ lib/index.ts
@@ -1,3 +1,0 @@
-export * from './src/components'
-export * from './src/res'
-export * from './src/screens'