+++ lib/src/components/Buttons/index.ts
@@ -1,4 +1,0 @@
-export { default as PrimaryButton } from './PrimaryButton'
-export { default as SecondaryButton } from './SecondaryButton'
-export { default as TertiaryButton, TertiaryButtonDisplayStyle } from './TertiaryButton'
-export * from './ButtonProps'