+++ src/components/Banners/BannerProps.ts
@@ -1,0 +1,12 @@
+import { ImageSourcePropType, ViewProps } from 'react-native'
+
+export enum BannerDisplayStyle {
+  Warning,
+  Error,
+}
+
+export interface BannerProps extends ViewProps {
+  displayStyle?: ImageSourcePropType
+  message: string
+  onPress?: () => void
+}