+++ src/components/Buttons/index.ts
@@ -1,0 +1,4 @@
+export { default as PrimaryButton } from './PrimaryButton'
+export { default as SecondaryButton } from './SecondaryButton'
+export { default as TertiaryButton } from './TertiaryButton'
+export * from './ButtonProps'